% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_ma_cpp}
\alias{compute_ma_cpp}
\title{Compute moving averages efficiently}
\usage{
compute_ma_cpp(x, windows)
}
\arguments{
\item{x}{Numeric vector}

\item{windows}{Integer vector of window sizes}
}
\value{
Matrix where each column is an MA with different window
}
\description{
Compute moving averages efficiently
}
