% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_rollminmax_cpp}
\alias{compute_rollminmax_cpp}
\title{Compute rolling min/max}
\usage{
compute_rollminmax_cpp(x, windows, compute_min = TRUE)
}
\arguments{
\item{x}{Numeric vector}

\item{windows}{Integer vector of window sizes}

\item{compute_min}{Logical, if TRUE compute min, else max}
}
\value{
Matrix where each column is rolling min or max
}
\description{
Compute rolling min/max
}
