% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_rollsd_cpp}
\alias{compute_rollsd_cpp}
\title{Compute rolling standard deviation}
\usage{
compute_rollsd_cpp(x, windows)
}
\arguments{
\item{x}{Numeric vector}

\item{windows}{Integer vector of window sizes}
}
\value{
Matrix where each column is a rolling SD
}
\description{
Compute rolling standard deviation
}
