% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_rollslope_cpp}
\alias{compute_rollslope_cpp}
\title{Compute rolling linear trend slope}
\usage{
compute_rollslope_cpp(x, windows)
}
\arguments{
\item{x}{Numeric vector}

\item{windows}{Integer vector of window sizes}
}
\value{
Matrix where each column is rolling slope
}
\description{
Compute rolling linear trend slope
}
