% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap_filling.R
\name{.fill_custom}
\alias{.fill_custom}
\title{Custom User Function}
\usage{
.fill_custom(y, dates, params = list())
}
\arguments{
\item{y}{Numeric vector}

\item{dates}{Date vector}

\item{params}{List with:
\itemize{
\item fn: User function with signature fn(y, dates, params)
}}
}
\value{
Filled numeric vector
}
\description{
Apply user-provided filling function.
}
\keyword{internal}
