% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{make_target_feats_cpp}
\alias{make_target_feats_cpp}
\title{Create target features for a single time point (used in recursive forecasting)}
\usage{
make_target_feats_cpp(
  y_hist,
  p = NULL,
  q = NULL,
  roll_windows = NULL,
  roll_stats = NULL,
  trend_windows = NULL
)
}
\arguments{
\item{y_hist}{Numeric vector of historical target values}

\item{p}{Integer, number of lags}

\item{q}{Integer vector of MA windows}

\item{roll_windows}{Integer vector of rolling stat windows}

\item{roll_stats}{Character vector of stats to compute}

\item{trend_windows}{Integer vector for rolling slopes}
}
\value{
Named list of feature values
}
\description{
Create target features for a single time point (used in recursive forecasting)
}
