% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_forecast.R
\name{print.cv_forecast}
\alias{print.cv_forecast}
\title{Print method for cv_forecast}
\usage{
\method{print}{cv_forecast}(x, ...)
}
\arguments{
\item{x}{A cv_forecast object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for cv_forecast
}
