% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{print.TimeSeries}
\alias{print.TimeSeries}
\title{Print method for TimeSeries}
\usage{
\method{print}{TimeSeries}(x, ...)
}
\arguments{
\item{x}{A TimeSeries object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for TimeSeries
}
