% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_forecast.R
\name{summary.cv_forecast}
\alias{summary.cv_forecast}
\title{Summary method for cv_forecast}
\usage{
\method{summary}{cv_forecast}(object, ...)
}
\arguments{
\item{object}{A cv_forecast object}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary method for cv_forecast
}
